refactor(db) never check for opened connection in PostgreSQL
With this connector, we will always reconnect to the same DB instance in
case a given connection is close during a migration step, so we don't
fear closing a connection while migrating, nor should we abort if we
find out that we ever did.

As such, this change prevents aborting migrations if we realize we
closed a connection through the migrations process, which results in a
more robust migration framework for this connector.