fix(schema) fix errors when encoding default empty arrays
Previously, a schema such as this:
```
{ type = "array",
  default = {},
  elements = { type = "string" },
}
```
would be encoded in json as a `{}` instead of as a `[]` (the Lua table given for the default value would be copied, and its metatable would remain)

This change sets the metatable to `cjson.array_mt` and also adds tests for other possible values given to the default (i.e. `cjson.empty_array`).